chore: refactor types
