+++ package.json
@@ -27,7 +27,6 @@
     "test:unit": "jest test/unit --forceExit"
   },
   "dependencies": {
-    "ufo": "^0.5.2",
     "allowlist": "^0.1.1",
     "consola": "^2.15.0",
     "defu": "^3.2.2",
@@ -37,6 +36,7 @@
     "is-https": "^3.0.2",
     "node-fetch": "^2.6.1",
     "requrl": "^3.0.2",
+    "ufo": "^0.5.2",
     "upath": "^2.0.1"
   },
   "devDependencies": {
