+++ src/generate.ts
@@ -3,9 +3,9 @@ import { createWriteStream, promises as fsp } from 'fs'
 import { promisify } from 'util'
 import stream from 'stream'
 import { join } from 'upath'
-import { ModuleOptions } from 'src/module'
 import fetch from 'node-fetch'
 import { joinURL, hasProtocol } from 'ufo'
+import { ModuleOptions } from './types'
 import { getFileExtension } from './runtime/utils'
 import { hash, logger } from './utils'
