+++ src/index.ts
@@ -1,3 +1,2 @@
 export * from './module'
 export * from './types'
-export type { ProviderFactory } from './provider'
