+++ src/runtime/sizes.ts
@@ -1,6 +1,5 @@
-import type { ImageSize, ImageOptions } from '../types'
+import type { ImageCTX, ImageSize, ImageOptions } from '../types'
 import { getFileExtension } from './utils'
-import type { ImageCTX } from './image'
 
 export type InputSizes = Partial<ImageSize>[] | string | boolean
