+++ src/types/global.ts
@@ -1,0 +1,17 @@
+import type { $Image } from './image'
+
+declare module '@nuxt/types' {
+  interface Context {
+    $img: $Image
+  }
+
+  interface NuxtAppOptions {
+    $img: $Image
+  }
+}
+
+declare module 'vue/types/vue' {
+  interface Vue {
+    $img: $Image
+  }
+}
