+++ src/types/index.ts
@@ -1,0 +1,5 @@
+import './global'
+
+export * from './image'
+export * from './module'
+export * from './provider'
