+++ tsconfig.json
@@ -13,7 +13,7 @@
       "@types/node"
     ],
     "paths": {
-      "@nuxt/image/*": [ "src/*" ],
+      "@nuxt/image/*": [ "src/*" ]
     }
   },
   "include": [
